poulstar/pyproject.toml
[tool.poetry]
name = "poulstar"
version = "0.1.0"
description = ""
authors = ["alirezaaraby <alirezaaraby5@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
poulstar = "poulstar.cli:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
